Refactor DB_PATH to allow env override
