+++ cyclone/cyclone_console/commands/alerts.py
@@ -4,7 +4,8 @@ import typer
 from rich.table import Table
 from rich.console import Console
 from cyclone.cyclone_alert_service import CycloneAlertService
-from core.core_imports import configure_console_log, DB_PATH
+from core.core_imports import configure_console_log
+from core.constants import DB_PATH
 from data.data_locker import DataLocker
 
 app = typer.Typer(help="🔔 Alert tools: view, enrich, evaluate")
