+++ cyclone/cyclone_console/commands/positions.py
@@ -3,7 +3,8 @@
 import typer
 from rich.table import Table
 from rich.console import Console
-from core.core_imports import configure_console_log, DB_PATH
+from core.core_imports import configure_console_log
+from core.constants import DB_PATH
 from data.data_locker import DataLocker
 from cyclone.cyclone_position_service import CyclonePositionService
 from core.logging import log
