+++ cyclone/cyclone_portfolio_service.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 from data.data_locker import DataLocker
 from data.alert import AlertType, Condition
 from alert_core.alert_utils import log_alert_summary
-from core.core_imports import DB_PATH
+from core.constants import DB_PATH
 from core.logging import log
