+++ cyclone/cyclone_report_generator.py
@@ -4,7 +4,7 @@ import datetime
 import requests
 import sqlite3
 from core.constants import LOG_DATE_FORMAT, LOG_DIR  # Import LOG_DIR
-from core.core_imports import DB_PATH
+from core.constants import DB_PATH
 
 def query_update_ledger():
     """
