+++ cyclone_app.py
@@ -6,7 +6,8 @@ from rich.prompt import Prompt
 from rich.table import Table
 from cyclone.cyclone_engine import Cyclone
 from monitor.monitor_core import MonitorCore
-from core.core_imports import configure_console_log, DB_PATH
+from core.core_imports import configure_console_log
+from core.constants import DB_PATH
 from data.data_locker import DataLocker
 
 console = Console()
