+++ data/insert_wallets.py
@@ -4,7 +4,8 @@ import sqlite3
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from core.core_imports import DB_PATH, configure_console_log
+from core.core_imports import configure_console_log
+from core.constants import DB_PATH
 from data.data_locker import DataLocker
 
 def ensure_wallets_table(db):
