+++ monitor/xcom_monitor.py
@@ -5,7 +5,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 from monitor.base_monitor import BaseMonitor
 from data.data_locker import DataLocker
 from xcom.xcom_core import XComCore
-from core.core_imports import DB_PATH
+from core.constants import DB_PATH
 from core.logging import log
